move database fetching to homepage js
